requirements.txt
pytorch
tensorboard
